requirements.txt
faker==24.8.0
psycopg2-binary
pymysql==1.1.0
pymongo==4.6.3